+++ R/manyenviron-memberships.R
@@ -6,8 +6,8 @@
 #' please use the `data_source()` and `data_contrast()` functions.
 #'\describe{
 #' \item{ECOLEX_MEM: }{A dataset with 25003 observations and the following
-#' 10 variables: manyID, CountryID, Title, Beg, End, SignatureCountry,
-#' Force, Rat, ecolexID, treatyID.}
+#' 10 variables: manyID, CountryID, Title, Beg, End, SignatureCountry, Force,
+#' Rat, ecolexID, treatyID.}
 #' \item{IEADB_MEM: }{A dataset with 15466 observations and the following
 #' 12 variables: manyID, CountryID, Title, Beg, End, SignatureCountry,
 #' Signature, Rat, Force, DocType, ieadbID, treatyID.}
@@ -18,9 +18,9 @@
 #' 31 variables: manyID, treatyID, CountryID, Title, Beg, End, Signature,
 #' Force, CountrySignature, CountryRat, CountryForce, CountryForce2,
 #' CountryForce3, Term, CountryWithdrawal, CountryWithdrawal2, gengID,
-#' ecolexID, ieaID, Comments, Deposit, obsolete, ProvisionalApp,
-#' Reservation, verified, Notes, CountryForce_ecolex, CountryForce_iea,
-#' Consent, Acceptance, Accession.}
+#' ecolexID, ieaID, Comments, Deposit, obsolete, ProvisionalApp, Reservation,
+#' verified, Notes, CountryForce_ecolex, CountryForce_iea, Consent, Acceptance,
+#' Accession.}
 #' }
 
 #'